fix(parts): extract _parts_shared helpers (CQ-002 step 1/4)
Closes #118 step 1. v2 teardown CQ-002.

`backend/app/api/routes/parts.py` is 1209 lines (4x the next-largest
router) and mixes CRUD, archive, bulk-delete, MPN-uniqueness, asset
serving, provider linking, custom-fields, scan-import, and bag-
signature lookup. The plan splits it into focused files; this PR is
the first step.

Lift the shared helpers and request schemas into a new
`_parts_shared.py`:
  - PartIn, PartPatch, BulkDeleteIn (Pydantic)
  - image_urls_for_parts, serialize_part, get_part

`parts.py` re-exports them under their existing private names
(_image_urls_for_parts, _serialize, _get_part) so call-site references
stay byte-identical and the diff is minimal. Subsequent step PRs
(parts_assets, parts_provider refresh, parts_bulk) import from
_parts_shared without further duplication.

No behaviour change. URL surface, response envelope, and dependency
chains preserved.

New tests/test_parts_routing.py snapshots every (method, path) under
/api/parts so the next split-step PRs can't accidentally drop or
rename an endpoint.

Remaining steps (deferred to follow-up PRs):
  - extract /assets/{ws_id}/{filename} into parts_assets.py
  - move refresh-from-provider into parts_provider.py
  - extract bulk-delete + bulk-import-from-scan + by-bag-signature +
    quick-remove-bag into parts_bulk.py
