+++ backend/app/api/routes/_parts_shared.py
@@ -1,0 +1,156 @@
+"""Shared helpers + Pydantic schemas for the parts router family.
+
+`backend/app/api/routes/parts.py` was 1209 lines (CQ-002) and growing —
+the next-largest router is 4× smaller. The plan in #118 splits it into
+focused files (parts CRUD, parts_assets, parts_provider, parts_bulk).
+This module is step 1 of the split: it lifts the helpers and request
+schemas that each split file would otherwise need to duplicate or
+re-import from a moving target.
+
+Nothing about behaviour changes here. Every function and schema below
+is an exact lift from `parts.py`. Subsequent PRs (#118-step-2..4) move
+endpoint groups out of `parts.py` and import from this module.
+"""
+from __future__ import annotations
+
+from typing import Literal
+from uuid import UUID
+
+from fastapi import HTTPException, status
+from pydantic import BaseModel, ConfigDict, Field
+from sqlalchemy import select
+
+from app.domain.custom_fields.models import CustomField
+from app.domain.parts.models import Part
+
+
+class PartIn(BaseModel):
+    model_config = ConfigDict(extra="forbid")
+
+    part_type: Literal["linked", "local", "meta", "sub_assembly"] = "local"
+    # Optional — defaults to mpn server-side when blank, so the operator can
+    # paste an MPN and skip the name field. At least one of name/mpn must be
+    # set; the create endpoint enforces that explicitly.
+    name: str | None = Field(default=None, max_length=300)
+    manufacturer: str | None = None
+    mpn: str | None = None
+    internal_part_number: str | None = None
+    description: str | None = None
+    notes_markdown: str | None = None
+    footprint: str | None = None
+    low_stock_report_quantity: int | None = None
+    attrition_percentage: float = 0
+    attrition_min_quantity: int = 0
+    default_storage_location_id: UUID | None = None
+    default_storage_mandatory: bool = False
+    serialized: bool = False
+
+
+class PartPatch(BaseModel):
+    model_config = ConfigDict(extra="forbid")
+
+    name: str | None = None
+    manufacturer: str | None = None
+    mpn: str | None = None
+    internal_part_number: str | None = None
+    description: str | None = None
+    notes_markdown: str | None = None
+    footprint: str | None = None
+    low_stock_report_quantity: int | None = None
+    attrition_percentage: float | None = None
+    attrition_min_quantity: int | None = None
+    default_storage_location_id: UUID | None = None
+    default_storage_mandatory: bool | None = None
+    serialized: bool | None = None
+    published: bool | None = None
+    # Command flag: when true, drops the provider link, clears
+    # last_refresh_at, resets description_locally_edited, and converts
+    # every {provider, override} custom_field row on this part to
+    # `manual` (override rows lose their original_value).
+    unlink_provider: bool | None = None
+
+
+class BulkDeleteIn(BaseModel):
+    model_config = ConfigDict(extra="forbid")
+
+    # 100 is plenty for a single-shot multi-select; if a user wants to wipe
+    # more they can run it twice. Keeps the response payload bounded.
+    part_ids: list[UUID] = Field(min_length=1, max_length=100)
+
+
+def image_urls_for_parts(db, ws_id, part_ids: list) -> dict:
+    """Single-shot SELECT for the per-part image_url custom_field row,
+    keyed by part_id. The /parts list endpoint uses this so we don't
+    fire one query per row."""
+    if not part_ids:
+        return {}
+    rows = db.execute(
+        select(CustomField.object_id, CustomField.value)
+        .where(CustomField.workspace_id == ws_id)
+        .where(CustomField.object_type == "part")
+        .where(CustomField.object_id.in_(part_ids))
+        .where(CustomField.key == "image_url")
+    ).all()
+    return {pid: val for pid, val in rows}
+
+
+def serialize_part(
+    p: Part,
+    *,
+    on_hand: int | None = None,
+    reserved: int | None = None,
+    available: int | None = None,
+    image_url: str | None = None,
+) -> dict:
+    if reserved is None:
+        reserved = 0
+    if available is None and on_hand is not None:
+        available = on_hand - reserved
+    return {
+        "id": str(p.id),
+        "part_type": p.part_type,
+        "name": p.name,
+        "manufacturer": p.manufacturer,
+        "mpn": p.mpn,
+        "internal_part_number": p.internal_part_number,
+        "description": p.description,
+        "footprint": p.footprint,
+        "notes_markdown": p.notes_markdown,
+        "low_stock_report_quantity": p.low_stock_report_quantity,
+        "attrition_percentage": float(p.attrition_percentage or 0),
+        "attrition_min_quantity": p.attrition_min_quantity or 0,
+        "default_storage_location_id": str(p.default_storage_location_id) if p.default_storage_location_id else None,
+        "default_storage_mandatory": p.default_storage_mandatory,
+        "serialized": p.serialized,
+        "published": bool(p.published),
+        "linked_provider": p.linked_provider,
+        "linked_external_id": p.linked_external_id,
+        "last_refresh_at": p.last_refresh_at.isoformat() if p.last_refresh_at else None,
+        "description_locally_edited": bool(p.description_locally_edited),
+        "archived_at": p.archived_at.isoformat() if p.archived_at else None,
+        "on_hand": on_hand,
+        "reserved": reserved,
+        "available": available if available is not None else 0,
+        # Sourced from the part's `image_url` custom_field — the post-
+        # download local path or a fallback to the upstream URL when the
+        # download failed. None when no image was ever attached.
+        "image_url": image_url,
+    }
+
+
+def get_part(db, ws_id, part_id, *, include_archived: bool = False) -> Part:
+    """Fetch a workspace-owned Part. Default refuses archived rows so
+    write paths (PATCH, substitute add/remove, BOM bind) can't bind
+    against a part the workspace already retired (BE2-016). Read-only
+    surfaces pass `include_archived=True` so the detail page and
+    activity timeline still load for archived parts.
+    """
+    p = db.get(Part, part_id)
+    if not p or p.workspace_id != ws_id:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="part not found")
+    if not include_archived and p.archived_at is not None:
+        # 404 (not 400) — the part is "not available" for binds. We
+        # don't distinguish "doesn't exist" from "archived" because the
+        # client treats both as "this id is dead".
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="part not found")
+    return p
